requirements.txt
python-can
paho-mqtt
pyyaml
